refactor: Member 패키지 내부 패키지 구조 변경
