+++ src/main/java/coupon/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package coupon.member;
+package coupon.member.repository;
 
+import coupon.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
